update scala-js-dom
- makes EmitterOps.valueAsNumber return Double instead of Int

+++ src/main/scala/outwatch/dom/helpers/EmitterOps.scala
@@ -10,14 +10,14 @@ trait EmitterOps {
     object target {
       def value: EmitterBuilder[E, String, Emitter] = builder.map(_.target.asInstanceOf[html.Input].value)
 
-      def valueAsNumber: EmitterBuilder[E, Int, Emitter] = builder.map(_.target.asInstanceOf[html.Input].valueAsNumber)
+      def valueAsNumber: EmitterBuilder[E, Double, Emitter] = builder.map(_.target.asInstanceOf[html.Input].valueAsNumber)
 
       def checked: EmitterBuilder[E, Boolean, Emitter] = builder.map(_.target.asInstanceOf[html.Input].checked)
     }
 
     def value: EmitterBuilder[E, String, Emitter] = builder.map(e => e.currentTarget.asInstanceOf[html.Input].value)
 
-    def valueAsNumber: EmitterBuilder[E, Int, Emitter] = builder.map(e => e.currentTarget.asInstanceOf[html.Input].valueAsNumber)
+    def valueAsNumber: EmitterBuilder[E, Double, Emitter] = builder.map(e => e.currentTarget.asInstanceOf[html.Input].valueAsNumber)
 
     def checked: EmitterBuilder[E, Boolean, Emitter] = builder.map(e => e.currentTarget.asInstanceOf[html.Input].checked)
